food-delivery-app/requirements.txt
Django>=3.2
djangorestframework
python-dotenv
Pillow
django-cors-headers